`bench`'s error messages are more specific
Thanks @cppforlife for the code review

Made (MB) be printed as floats because otherwise there are small
discrepancies when manually dividing the MB by the seconds.  and
comparing to the calculated MB/s.

+++ gobonniego/gobonniego.go
@@ -60,7 +60,7 @@ func main() {
 	bm.RunSequentialWriteTest()
 	if verbose {
 		log.Printf("Written (MiB): %d\n", bm.Result.WrittenBytes>>20)
-		log.Printf("Written (MB): %d\n", bm.Result.WrittenBytes/1000000)
+		log.Printf("Written (MB): %f\n", float64(bm.Result.WrittenBytes)/1000000)
 		log.Printf("Duration (seconds): %f\n", bm.Result.WrittenDuration.Seconds())
 	}
 	fmt.Printf("Sequential Write MB/s: %0.2f\n",
@@ -69,7 +69,7 @@ func main() {
 	bm.RunSequentialReadTest()
 	if verbose {
 		log.Printf("Read (MiB): %d\n", bm.Result.ReadBytes>>20)
-		log.Printf("Read (MB): %d\n", bm.Result.ReadBytes/1000000)
+		log.Printf("Read (MB): %f\n", float64(bm.Result.ReadBytes)/1000000)
 		log.Printf("Duration (seconds): %f\n", bm.Result.ReadDuration.Seconds())
 	}
 	fmt.Printf("Sequential Read MB/s: %0.2f\n",
